tools/test_runner/requirements.txt
bleak>=0.22.0
PyYAML>=6.0.0
rich>=13.7.0
pytest>=8.0.0
